+++ src/main/java/com/cloudogu/scm/teamscale/pullrequest/PullRequestLinkEnricher.java
@@ -1,0 +1,73 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2020-present Cloudogu GmbH and Contributors
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package com.cloudogu.scm.teamscale.pullrequest;
+
+import com.cloudogu.scm.review.pullrequest.service.PullRequest;
+import sonia.scm.api.v2.resources.Enrich;
+import sonia.scm.api.v2.resources.HalAppender;
+import sonia.scm.api.v2.resources.HalEnricher;
+import sonia.scm.api.v2.resources.HalEnricherContext;
+import sonia.scm.api.v2.resources.LinkBuilder;
+import sonia.scm.api.v2.resources.ScmPathInfoStore;
+import sonia.scm.plugin.Extension;
+import sonia.scm.plugin.Requires;
+import sonia.scm.repository.Repository;
+import sonia.scm.repository.RepositoryPermissions;
+
+import javax.inject.Inject;
+import javax.inject.Provider;
+
+import static com.cloudogu.scm.teamscale.Constants.READ_FINDINGS_PERMISSION;
+
+@Requires("scm-review-plugin")
+@Enrich(PullRequest.class)
+@Extension
+public class PullRequestLinkEnricher implements HalEnricher {
+
+  private final Provider<ScmPathInfoStore> pathInfoStore;
+
+  @Inject
+  public PullRequestLinkEnricher(Provider<ScmPathInfoStore> pathInfoStore) {
+    this.pathInfoStore = pathInfoStore;
+  }
+
+  @Override
+  public void enrich(HalEnricherContext context, HalAppender appender) {
+
+    Repository repository = context.oneRequireByType(Repository.class);
+    PullRequest pullRequest = context.oneRequireByType(PullRequest.class);
+
+    if (RepositoryPermissions.custom(READ_FINDINGS_PERMISSION, repository).isPermitted()) {
+      appender.appendLink("teamscaleFindings", createFindingsLink(repository, pullRequest.getId()));
+    }
+  }
+
+  private String createFindingsLink(Repository repository, String pullRequestId) {
+    LinkBuilder linkBuilder = new LinkBuilder(pathInfoStore.get().get(), PullRequestResource.class);
+    return linkBuilder
+      .method("getFindings")
+      .parameters(repository.getNamespace(), repository.getName(), pullRequestId)
+      .href();
+  }
+}