+++ src/test/java/com/cloudogu/scm/teamscale/pullrequest/PullRequestResourceTest.java
@@ -27,24 +27,34 @@ import com.cloudogu.scm.review.comment.api.CommentResource;
 import com.cloudogu.scm.review.comment.api.CommentRootResource;
 import com.cloudogu.scm.review.pullrequest.api.PullRequestRootResource;
 import com.cloudogu.scm.review.pullrequest.api.PullRequestSelector;
+import org.apache.shiro.authz.AuthorizationException;
+import org.apache.shiro.subject.Subject;
+import org.apache.shiro.util.ThreadContext;
 import org.jboss.resteasy.mock.MockHttpRequest;
 import org.jboss.resteasy.mock.MockHttpResponse;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import sonia.scm.repository.Repository;
+import sonia.scm.repository.RepositoryManager;
 import sonia.scm.repository.RepositoryTestData;
 import sonia.scm.web.RestDispatcher;
 import sonia.scm.web.VndMediaType;
 
 import javax.ws.rs.core.UriInfo;
+import java.io.UnsupportedEncodingException;
 import java.net.URISyntaxException;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -53,7 +63,8 @@ public class PullRequestResourceTest {
 
   private final Repository REPOSITORY = RepositoryTestData.createHeartOfGold();
 
-  private static final String MEDIATYPE = VndMediaType.PREFIX + "teamscale_pr" + VndMediaType.SUFFIX;
+  private static final String PR_MEDIATYPE = VndMediaType.PREFIX + "teamscalePullRequest" + VndMediaType.SUFFIX;
+  private static final String FINDINGS_MEDIATYPE = VndMediaType.PREFIX + "teamscaleFindings" + VndMediaType.SUFFIX;
 
   @Mock
   private PullRequestRootResource pullRequestRootResource;
@@ -62,16 +73,28 @@ public class PullRequestResourceTest {
   private com.cloudogu.scm.review.pullrequest.api.PullRequestResource reviewPullRequestResource;
 
   @Mock
+  private FindingsService findingsService;
+
+  @Mock
+  private RepositoryManager repositoryManager;
+
+  @Mock
+  private FindingsMapper findingsMapper;
+
+  @Mock
   private CommentRootResource commentRootResource;
 
   @Mock
   private CommentResource commentResource;
 
+  @Mock
+  private Subject subject;
+
   private RestDispatcher restDispatcher;
 
   @BeforeEach
   void initDispatcher() {
-    PullRequestResource pullRequestResource = new PullRequestResource(pullRequestRootResource);
+    PullRequestResource pullRequestResource = new PullRequestResource(pullRequestRootResource, repositoryManager, findingsService, findingsMapper);
     restDispatcher = new RestDispatcher();
     restDispatcher.addSingletonResource(pullRequestResource);
   }
@@ -80,7 +103,7 @@ public class PullRequestResourceTest {
   void shouldGetAllPullRequests() throws URISyntaxException {
     MockHttpRequest request = MockHttpRequest
       .get("/v2/teamscale/pull-request/" + REPOSITORY.getNamespace() + "/" + REPOSITORY.getName())
-      .contentType(MEDIATYPE);
+      .contentType(PR_MEDIATYPE);
 
     MockHttpResponse response = new MockHttpResponse();
 
@@ -94,7 +117,7 @@ public class PullRequestResourceTest {
     when(pullRequestRootResource.getPullRequestResource()).thenReturn(reviewPullRequestResource);
     MockHttpRequest request = MockHttpRequest
       .get("/v2/teamscale/pull-request/" + REPOSITORY.getNamespace() + "/" + REPOSITORY.getName() + "/1")
-      .contentType(MEDIATYPE);
+      .contentType(PR_MEDIATYPE);
 
     MockHttpResponse response = new MockHttpResponse();
 
@@ -110,7 +133,7 @@ public class PullRequestResourceTest {
     when(reviewPullRequestResource.comments()).thenReturn(commentRootResource);
     MockHttpRequest request = MockHttpRequest
       .get("/v2/teamscale/pull-request/comments/" + REPOSITORY.getNamespace() + "/" + REPOSITORY.getName() + "/1")
-      .contentType(MEDIATYPE);
+      .contentType(PR_MEDIATYPE);
 
     MockHttpResponse response = new MockHttpResponse();
 
@@ -128,7 +151,7 @@ public class PullRequestResourceTest {
     when(commentRootResource.getCommentResource()).thenReturn(commentResource);
     MockHttpRequest request = MockHttpRequest
       .delete("/v2/teamscale/pull-request/comments/" + REPOSITORY.getNamespace() + "/" + REPOSITORY.getName() + "/1/abc")
-      .contentType(MEDIATYPE);
+      .contentType(PR_MEDIATYPE);
 
     MockHttpResponse response = new MockHttpResponse();
 
@@ -149,9 +172,9 @@ public class PullRequestResourceTest {
     byte[] commentJson = "{\"comment\" : \"this is my comment\"}".getBytes();
 
     MockHttpRequest request = MockHttpRequest
-      .post("/v2/teamscale/pull-request/comments/" + REPOSITORY.getNamespace() + "/" + REPOSITORY.getName() + "/1/abc?sourceRevision=source&targetRevision=target")
+      .post("/v2/teamscale/pull-request/comments/" + REPOSITORY.getNamespace() + "/" + REPOSITORY.getName() + "/1?sourceRevision=source&targetRevision=target")
       .content(commentJson)
-      .contentType(MEDIATYPE);
+      .contentType(PR_MEDIATYPE);
 
     MockHttpResponse response = new MockHttpResponse();
 
@@ -163,4 +186,98 @@ public class PullRequestResourceTest {
     assertThat(response.getStatus()).isEqualTo(204);
   }
 
+  @Nested
+  class WithSubject {
+
+    @BeforeEach
+    void bindSubject() {
+      ThreadContext.bind(subject);
+    }
+
+    @AfterEach
+    void tearDownSubject() {
+      ThreadContext.unbindSubject();
+    }
+
+    @Test
+    void shouldThrowAuthorizationExceptionIfNotPermittedToRead() throws URISyntaxException {
+      String permission = "repository:readTeamscaleFindings:" + REPOSITORY.getId();
+      doThrow(AuthorizationException.class).when(subject).checkPermission(permission);
+
+      when(repositoryManager.get(REPOSITORY.getNamespaceAndName())).thenReturn(REPOSITORY);
+
+      MockHttpRequest request = MockHttpRequest
+        .get("/v2/teamscale/pull-request/" + REPOSITORY.getNamespace() + "/" + REPOSITORY.getName() + "/1/findings")
+        .contentType(FINDINGS_MEDIATYPE);
+
+      MockHttpResponse response = new MockHttpResponse();
+
+      restDispatcher.invoke(request, response);
+
+      assertThat(response.getStatus()).isEqualTo(403);
+      verify(findingsService, never()).getFindings(any(), any());
+    }
+
+    @Test
+    void shouldThrowAuthorizationExceptionIfNotPermittedToWrite() throws URISyntaxException {
+      String permission = "repository:writeTeamscaleFindings:" + REPOSITORY.getId();
+      doThrow(AuthorizationException.class).when(subject).checkPermission(permission);
+
+      String content = "teamscale findings: 2";
+      byte[] contentJson = ("{\"content\" : \"" + content + "\"}").getBytes();
+      when(repositoryManager.get(REPOSITORY.getNamespaceAndName())).thenReturn(REPOSITORY);
+
+      MockHttpRequest request = MockHttpRequest
+        .put("/v2/teamscale/pull-request/" + REPOSITORY.getNamespace() + "/" + REPOSITORY.getName() + "/1/findings")
+        .content(contentJson)
+        .contentType(FINDINGS_MEDIATYPE);
+
+      MockHttpResponse response = new MockHttpResponse();
+
+      restDispatcher.invoke(request, response);
+
+      assertThat(response.getStatus()).isEqualTo(403);
+      verify(findingsService, never()).setFindings(any(), any(), any());
+    }
+
+    @Test
+    void shouldGetFindings() throws URISyntaxException, UnsupportedEncodingException {
+      String content = "teamscale findings: 2";
+      Findings findings = new Findings(content);
+      when(findingsService.getFindings(REPOSITORY, "1")).thenReturn(findings);
+      when(repositoryManager.get(REPOSITORY.getNamespaceAndName())).thenReturn(REPOSITORY);
+      when(findingsMapper.map(findings, REPOSITORY, "1")).thenReturn(new FindingsDto(content));
+
+      MockHttpRequest request = MockHttpRequest
+        .get("/v2/teamscale/pull-request/" + REPOSITORY.getNamespace() + "/" + REPOSITORY.getName() + "/1/findings")
+        .contentType(FINDINGS_MEDIATYPE);
+
+      MockHttpResponse response = new MockHttpResponse();
+
+      restDispatcher.invoke(request, response);
+
+      assertThat(response.getStatus()).isEqualTo(200);
+      assertThat(response.getContentAsString()).isEqualTo("{\"content\":\"teamscale findings: 2\"}");
+    }
+
+    @Test
+    void shouldUpdateFindings() throws URISyntaxException {
+      String content = "teamscale findings: 2";
+      byte[] contentJson = ("{\"content\" : \"" + content + "\"}").getBytes();
+      when(repositoryManager.get(REPOSITORY.getNamespaceAndName())).thenReturn(REPOSITORY);
+
+      MockHttpRequest request = MockHttpRequest
+        .put("/v2/teamscale/pull-request/" + REPOSITORY.getNamespace() + "/" + REPOSITORY.getName() + "/1/findings")
+        .content(contentJson)
+        .contentType(FINDINGS_MEDIATYPE);
+
+      MockHttpResponse response = new MockHttpResponse();
+
+      restDispatcher.invoke(request, response);
+
+      verify(findingsService).setFindings(REPOSITORY, "1", content);
+      assertThat(response.getStatus()).isEqualTo(204);
+    }
+  }
 }
+